+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/MappingHandler.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
