+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -28,8 +28,8 @@ class AnnotationHandlerMappingTest {
         when(request.getRequestURI()).thenReturn("/get-test");
         when(request.getMethod()).thenReturn("GET");
 
-        final var handlerExecution = (HandlerExecution) handlerMapping.getHandler(request);
-        final var modelAndView = handlerExecution.handle(request, response);
+        final var handlerExecution = handlerMapping.getHandler(request);
+        final var modelAndView = handlerExecution.handle(request);
 
         assertThat(modelAndView.getObject("id")).isEqualTo("gugu");
     }
@@ -43,8 +43,8 @@ class AnnotationHandlerMappingTest {
         when(request.getRequestURI()).thenReturn("/post-test");
         when(request.getMethod()).thenReturn("POST");
 
-        final var handlerExecution = (HandlerExecution) handlerMapping.getHandler(request);
-        final var modelAndView = handlerExecution.handle(request, response);
+        final var handlerExecution = handlerMapping.getHandler(request);
+        final var modelAndView = handlerExecution.handle(request);
 
         assertThat(modelAndView.getObject("id")).isEqualTo("gugu");
     }
